Fixed Makefile and compilation error
