+++ dos-madm/.editorconfig
@@ -1,0 +1,3 @@
+[Makefile]
+indent_style = tab
+
