+++ dos-madm/.gitignore
@@ -1,0 +1,2 @@
+*.o
+*.exe
