+++ dos-madm/ARITH.C
