+++ dos-madm/CURSOR.C
