+++ dos-madm/EDIT.C
@@ -8,6 +8,45 @@
 #include "keyboard.h"
 #include "display.h"
 #include "proto.h"
+//#include "conio_shim.h"
+
+//Hack(anita): Move conio stuff to the top of the file and move it out of madm.h
+#include <stdio.h>
+
+#if defined(__APPLE__) || defined(__linux__)
+
+#include <unistd.h>
+#include <termios.h>
+
+char getch(void)
+{
+    char buf = 0;
+    struct termios old = {0};
+    fflush(stdout);
+    if(tcgetattr(0, &old) < 0)
+        perror("tcsetattr()");
+    old.c_lflag &= ~ICANON;
+    old.c_lflag &= ~ECHO;
+    old.c_cc[VMIN] = 1;
+    old.c_cc[VTIME] = 0;
+    if(tcsetattr(0, TCSANOW, &old) < 0)
+        perror("tcsetattr ICANON");
+    if(read(0, &buf, 1) < 0)
+        perror("read()");
+    old.c_lflag |= ICANON;
+    old.c_lflag |= ECHO;
+    if(tcsetattr(0, TCSADRAIN, &old) < 0)
+        perror("tcsetattr ~ICANON");
+    printf("%c\n", buf);
+    return buf;
+}
+
+#else
+
+#include <conio.h>
+
+#endif
+
 
 /*
  * edit -- get commands from the keyboard and obey them
