+++ dos-madm/EXECUTE.C
@@ -6,6 +6,12 @@
 #include "madm.h"
 #include "proto.h"
 #include "keyboard.h"
+#include "stdbool.h"
+
+//Fixme(anita): This needs to be implemented - 1/28/2024
+_Bool kbhit(void) {
+	return 0;
+}
 
 void
 execute()
@@ -15,3 +21,4 @@ execute()
 		exec_instruction();
 	} while (MADM_status == RUNNING && !cmd_ready());
 }
+
