+++ dos-madm/EXEC_INS.C
