+++ dos-madm/FETCH.C
