+++ dos-madm/INIT.C
