+++ dos-madm/MADM.C
