+++ dos-madm/MADM.H
@@ -1,3 +1,6 @@
+#ifndef MADM_H
+#define MADM_H
+
 /*
  * madm.h -- definitions relating to the operating characteristics of
  *			 the prototype Manchester Mark I
@@ -55,37 +58,4 @@ typedef struct {
 extern LINE  Accumulator[], Control[], Store[];
 extern INSTRUCTION Staticisor;
 
-
-#if defined(LINUX) || defined(APPLE)
-
-
-#include <unistd.h>
-#include <termios.h>
-
-char getch(void)
-{
-    char buf = 0;
-    struct termios old = {0};
-    fflush(stdout);
-    if(tcgetattr(0, &old) < 0)
-        perror("tcsetattr()");
-    old.c_lflag &= ~ICANON;
-    old.c_lflag &= ~ECHO;
-    old.c_cc[VMIN] = 1;
-    old.c_cc[VTIME] = 0;
-    if(tcsetattr(0, TCSANOW, &old) < 0)
-        perror("tcsetattr ICANON");
-    if(read(0, &buf, 1) < 0)
-        perror("read()");
-    old.c_lflag |= ICANON;
-    old.c_lflag |= ECHO;
-    if(tcsetattr(0, TCSADRAIN, &old) < 0)
-        perror("tcsetattr ~ICANON");
-    printf("%c\n", buf);
-    return buf;
- }
-#else
-
-#include <conio.h>
-
-#endif
+#endif /* MADM_H */
