+++ dos-madm/MAKEFILE
@@ -2,11 +2,11 @@ CC=gcc
 LFLAGS=-g
 LIBS=
 SRC= arith.c control.c cursor.c display.c edit.c exec_ins.c \
-     execute.c fetch.c graphics.c init.c madm.c main.c  \
-     memory.c
+	execute.c fetch.c graphics.c init.c madm.c main.c  \
+	memory.c
 OBJS=arith.o control.o cursor.o display.o edit.o exec_ins.o \
-     execute.o fetch.o graphics.o init.o madm.o main.o   \
-     memory.o
+	execute.o fetch.o graphics.o init.o madm.o main.o   \
+	memory.o
 
 madm.exe: $(OBJS)
 	gcc $(LFLAGS) -o $@ $(OBJS) $(LIBS)
@@ -22,3 +22,6 @@ fetch.o graphics.o init.o memory.o: display.h
 cursor.o display.o graphics.o: graphics.h
 
 edit.o execute.o: keyboard.h
+
+clean:
+	rm *.o madm.exe
