+++ dos-madm/MEMORY.C
